requirements.txt
## 14. Requirements File


# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
django-cors-headers==4.3.0
drf-spectacular==0.26.5

# Database
psycopg2-binary==2.9.9

# Celery
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Channels (WebSockets)
channels==4.0.0
channels-redis==4.1.0

# External APIs
ccxt==4.1.47
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.26.2

yfinance==0.2.66

# Utilities
python-decouple==3.8
pillow==10.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
daphne==4.0.0
whitenoise==6.6.0